Shooting gone, see description
We all came up with a new idea, instead of focusing around the
platforming and shooting, we're going to focus on the act of throwing
objects at enemies, putting them on buttons and so on. So the weapon
system is gone, and you can now grab objects tagged with "grabbable".